requirements.txt
annotated-types==0.7.0
attrs==23.2.0
Automat==22.10.0
Babel==2.10.3
bcrypt==3.2.2
blinker==1.9.0
cachetools==6.2.2
certifi==2023.11.17
chardet==5.2.0
click==8.1.6
cloud-init==25.2
colorama==0.4.6
command-not-found==0.3
configobj==5.0.8
constantly==23.10.4
cryptography==41.0.7
dbus-python==1.3.2
distro==1.9.0
distro-info==1.7+build1
Flask==3.1.2
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
gyp==0.1
httplib2==0.20.4
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
itsdangerous==2.2.0
Jinja2==3.1.2
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.10.3
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
netifaces==0.11.0
oauthlib==3.2.2
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycurl==7.45.3
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.17.2
PyGObject==3.48.2
PyHamcrest==2.1.0
PyJWT==2.7.0
pyOpenSSL==23.2.0
pyparsing==3.1.1
pypdf==6.4.1
pyrsistent==0.20.0
pyserial==3.5
python-apt==2.7.7+ubuntu5
python-dotenv==1.2.1
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
rich==13.7.1
rsa==4.9.1
service-identity==24.1.0
setuptools==68.1.2
six==1.16.0
systemd-python==235
tqdm==4.67.1
Twisted==24.3.0
typing-inspection==0.4.2
typing_extensions==4.15.0
ubuntu-pro-client==8001
unattended-upgrades==0.1
uritemplate==4.2.0
urllib3==2.0.7
wadllib==1.3.6
Werkzeug==3.1.4
wheel==0.42.0
zope.interface==6.1
